int to long

+++ src/User.java
@@ -1,7 +1,7 @@
 
 public class User {
 	String name,address;
-	int phone_number;
+	long phone_number;
 	public String getName() {
 		return name;
 	}
@@ -14,10 +14,10 @@ public class User {
 	public void setAddress(String address) {
 		this.address = address;
 	}
-	public int getPhone_number() {
+	public long getPhone_number() {
 		return phone_number;
 	}
-	public void setPhone_number(int phone_number) {
+	public void setPhone_number(long phone_number) {
 		this.phone_number = phone_number;
 	}
 	public Order getCurrent_order() {
